Sounds Converted and implemented,Image Changes, Scenes with Racoon mascot and background, Teacher register, login-student, student register-login
Converted m4a sound file to mp3 to make them work with unity.
Removed background form the images.
Added the mascot Racoon and background to the scenes, implemented sounds when introducing the elements and in the tests.
Teachers can now register themselves and login. New students can now be registered to the database by the teachers.
Teachers can now see the statistics of the tests.
Teachers can now choose the number of a student and press play, the tests that have been completed will be recorded to the database with the chosen student's number.